+++ src/features/board/boardDetail/api/types/index.ts
